Refactoring all code
