+++ parsers/__init__.py
@@ -1,0 +1,2 @@
+from . import groups_parser
+from . import schedule_parser
