+++ rea_schedule_site/urls.py
@@ -1,0 +1,8 @@
+from django.conf.urls import include, url
+from django.contrib import admin
+from django.urls import path
+
+urlpatterns = [
+    path('admin/', admin.site.urls),
+    url(r'^', include('django_telegrambot.urls')),
+]
