+++ telegram_bot/schedule_wrappers.py
@@ -5,20 +5,26 @@ def create_schedule_markups(schedule):
     markups = []
     for day in schedule:
         buttons = []
-        buttons.append([InlineKeyboardButton(
-            day,
-            callback_data='#'
-        ), ])
+        buttons.append(
+            [
+                InlineKeyboardButton(
+                    day,
+                    callback_data='#'
+                ),
+            ]
+        )
         for lesson in schedule[day]:
             lesson_markup_text = schedule[day][lesson] or '---'
-            buttons.append([
-                InlineKeyboardButton(
-                    lesson,
-                    callback_data='#'),
-                InlineKeyboardButton(
-                    lesson_markup_text,
-                    callback_data='#'),
-            ])
+            buttons.append(
+                [
+                    InlineKeyboardButton(
+                        lesson,
+                        callback_data='#'),
+                    InlineKeyboardButton(
+                        lesson_markup_text,
+                        callback_data='#'),
+                ]
+            )
         markup = InlineKeyboardMarkup(buttons)
         markups.append(markup)
     return markups
@@ -54,3 +60,14 @@ def wrap_schedule_week(schedule):
         wrapped_day = wrap_schedule_with_ascii_lines(day)
         wrapped_schedule.append(wrapped_day)
     return wrapped_schedule
+
+
+def create_buttons_markup(options):
+    buttons = []
+    for option in options:
+        buttons.append([InlineKeyboardButton(
+            option,
+            callback_data=options[option]
+        ), ])
+    reply_markup = InlineKeyboardMarkup(buttons)
+    return reply_markup
